Move atmospheric modifier tests to own test module
Mode tests for modifiers in modifiers/atmosphere.py to their own test
module in tests/modifier_tests/test_atmosphere.py.
